Refactor Kubernetes client configuration to support in-cluster and local kubeconfig; update Docker Compose for email worker; enhance README with deployment instructions; update API server and container service images; add service account for API access; improve TypeScript definitions for gRPC services and deployment requests/responses; adjust TypeScript configuration for better module resolution.
