+++ k8s/db/deployment.yaml
@@ -17,7 +17,7 @@ spec:
     spec:
       containers:
         - name: scs-cloud-db
-          image: mongo:8.0.15-noble
+          image: mongo:latest
           ports:
             - containerPort: 27017
               name: scs-cloud-db
