+++ scsApiServer/.dockerignore
@@ -1,0 +1,7 @@
+node_modules
+npm-debug.log
+dist
+.git
+.vscode
+Dockerfile*
+README_OLD.md
