+++ scsApiServer/Dockerfile
@@ -1,19 +1,25 @@
-FROM node:22-alpine AS builder
-
-WORKDIR /build
+FROM node:22-alpine AS build
+WORKDIR /app
 
 COPY package*.json ./
 RUN npm install
-COPY containerService.proto containerService.proto
-COPY src/ src/
-COPY pb/ pb/
-COPY tsconfig.json tsconfig.json
+
+COPY tsconfig.json ./
+COPY containerService.proto ./
+COPY src ./src
+
 RUN npm run build
 
-FROM node:22-alpine AS runner
+RUN npm prune --production
 
+FROM node:22-alpine AS runner
 WORKDIR /app
-COPY --from=builder build/package*.json .
-COPY --from=builder build/node_modules node_modules
-COPY --from=builder build/dist dist/
-CMD ["npm","start"]
+ENV NODE_ENV=production
+
+COPY --from=build /app/package*.json ./
+COPY --from=build /app/node_modules ./node_modules
+COPY --from=build /app/dist ./dist
+COPY --from=build /app/containerService.proto ./
+
+EXPOSE 3000
+CMD ["node","dist/index.js"]
