+++ scsApiServer/src/pb/pb/DeleteDeploymentResponse.ts
