+++ scsApiServer/src/services/gRPC/gRPC.service.ts
@@ -1,9 +1,12 @@
 import path from 'path';
 import * as grpc from '@grpc/grpc-js';
 import * as protoLoader from '@grpc/proto-loader';
-import { ProtoGrpcType } from '../../../pb/containerService';
+import { ProtoGrpcType } from '../../pb/containerService';
 import "dotenv/config";
-import { Deployment } from '../../../pb/containerService/Deployment';
+import { Deployment } from '../../pb/containerService/Deployment';
+import { GetDeploymentsResponse__Output } from '../../pb/containerService/GetDeploymentsResponse';
+import { DeleteDeploymentResponse__Output } from '../../pb/containerService/DeleteDeploymentResponse';
+import { PostDeploymentResponse__Output } from '../../pb/containerService/PostDeploymentResponse';
 
 const PORT= process.env.CONTAINER_SERVICE_PORT || 8000;
 const PROTO_PATH = path.join(__dirname, '..','..','..','containerService.proto');
@@ -28,7 +31,7 @@ export const grpcClient= new grpcObject.pb.ContainerService(
 const deadline = new Date();
 deadline.setSeconds(deadline.getSeconds() + 5);
 
-grpcClient.waitForReady(deadline, (error) => {
+grpcClient.waitForReady(deadline, (error?: Error) => {
     if (error) {
         console.error('gRPC client connection error:', error);
     } else {
@@ -38,7 +41,7 @@ grpcClient.waitForReady(deadline, (error) => {
 
 export function GetDeployments(userId: string):Promise<Deployment[]|undefined> {
     return new Promise((resolve, reject) => {
-        grpcClient.getDeployments({userId: userId}, (error, response) => {
+        grpcClient.getDeployments({userId: userId}, (error: grpc.ServiceError | null, response: GetDeploymentsResponse__Output | undefined) => {
             if (error) {
                 reject(error);
             } else {
@@ -76,7 +79,7 @@ export function DeployContainer(userId: string, deploymentConfig: Deployment ):P
             createdAt: new Date().toISOString()
         };
 
-        grpcClient.CreateDeployment({ userId: userId, deployment: deploymentPayload }, (error, response) => {
+        grpcClient.CreateDeployment({ userId: userId, deployment: deploymentPayload }, (error: grpc.ServiceError | null, response: PostDeploymentResponse__Output | undefined) => {
             if (error) {
                 reject(error);
             } else {
@@ -88,7 +91,7 @@ export function DeployContainer(userId: string, deploymentConfig: Deployment ):P
 
 export function DeleteDeployment(deploymentId: string) {
     return new Promise((resolve, reject) => {
-        grpcClient.DeleteDeployment({ id: deploymentId}, (error, response) => {
+        grpcClient.DeleteDeployment({ id: deploymentId}, (error: grpc.ServiceError | null, response: DeleteDeploymentResponse__Output | undefined) => {
             if (error) {
                 reject(error);
             } else {
